pyproject.toml
[project]
name = "hyperbolic-rl"
version = "0.1.0"
description = "Reinforcement Learning with hyperbolic representations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gym==0.23.1",
    "gymnasium[other]==1.0.0",
    "kaleido>=1.0.0",
    "numpy>=1.24.4",
    "opencv-python>=4.8.1.78",
    "optuna-integration>=4.2.1",
    "optuna>=4.2.0",
    "pandas>=2.3.3",
    "plotly[express]>=6.3.0",
    "pytest>=8.3.3",
    "ruff>=0.13.2",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "swig>=4.2.1.post0",
    "tensorboard>=2.14.0",
    "threadpoolctl>=3.5.0",
    "torch-tb-profiler>=0.4.3",
    "torchinfo>=1.8.0",
    "tqdm>=4.66.6",
    "tyro>=0.9.32",
    "wandb>=0.22.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "procgen-mirror>=0.10.7",
    "minigrid>=3.0.0",
    "tensordict>=0.10.0",
    "ipdb>=0.13.13",
    "statsmodels>=0.14.4",
    "hydra-core>=1.3.2",
    "hydra-joblib-launcher>=1.2.0",
    "ale-py==0.10.2",
    "ipykernel>=6.30.1",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

# [tool.uv]
# constraint-dependencies = ["kaleido!=0.2.1.post1"]

[tool.ruff]
# See https://beta.ruff.rs/docs/rules/
lint.select = ["E", "F", "B", "UP", "C90", "RUF"]
# B028: Ignore explicit stacklevel`
# RUF013: Too many false positives (implicit optional)
# UP007: Changes Union to | which breaks pyrallis
lint.ignore = ["B028", "RUF013", "UP007"]

# Allow unused variables when underscore-prefixed.
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

lint.mccabe.max-complexity = 15

# Exclude a variety of commonly ignored directories.
exclude = [
    "*.egg-info",
    "__pycache__",
    ".eggs",
    ".git",
    "venv",
    "wandb",
    "log/",
    "runs/",
    "build/",
    ".ruff_cache",
    ".mypy_cache",
    "src/hyperbolic_math"
]

# Same as Black.
line-length = 127

# Assume Python 3.8
target-version = "py312"

[tool.black]
line-length = 127

[tool.isort]
profile = "black"
line_length = 127

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -vv"
testpaths = ["src/hyperbolic_math/tests"]
